GPT_related/requirements.txt
gradio==4.31.5
openai==1.30.4
PyPDF2==3.0.1
python-dotenv==1.0.1
